+++ basic/app/src/main/scala/freestyle/cassandra/sample/LowLevelApi.scala
@@ -19,64 +19,67 @@ package freestyle.cassandra.sample
 import java.util.UUID
 
 import cats.instances.future._
-import com.datastax.driver.core._
+import com.datastax.driver.core.{BoundStatement, PreparedStatement, Session}
 import freestyle._
 import freestyle.implicits._
-import freestyle.cassandra.api.{ClusterAPI, SessionAPI}
+import freestyle.cassandra.api.ClusterAPI
 import freestyle.cassandra.codecs._
-import freestyle.cassandra.query.interpolator.RuntimeCQLInterpolator._
-import freestyle.cassandra.query.interpolator._
 import freestyle.cassandra.sample.Implicits._
 import freestyle.cassandra.sample.Model._
 import freestyle.loggingJVM.implicits._
-import monix.cats._
 import monix.eval.{Task => MonixTask}
 import monix.execution.Scheduler
 
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
 
-
-object StringInterpolator extends App {
+object LowLevelApi extends App {
 
   import Modules.CassandraApp
-
-  implicit val executionContext: Scheduler = Scheduler.Implicits.global
 
   def connect[F[_]](implicit clusterAPI: ClusterAPI[F]): FreeS[F, Session] =
     clusterAPI.connectKeyspace("demodb")
 
-  def closeSession[F[_]](implicit sessionAPI: SessionAPI[F]): FreeS[F, Unit] = sessionAPI.close
+  def program[F[_]](implicit app: CassandraApp[F]): FreeS[F, User] = {
+
+    import app._
+
+    val newUser = User(UUID.randomUUID(), "Username")
+
+    def bindValues(st: PreparedStatement)(
+        implicit c1: ByteBufferCodec[UUID],
+        c2: ByteBufferCodec[String]): FreeS[F, BoundStatement] =
+      List(
+        statementAPI.setValueByName(_: BoundStatement, "id", newUser.id, c1),
+        statementAPI.setValueByName(_: BoundStatement, "name", newUser.name, c2))
+        .foldLeft[FreeS[F, BoundStatement]](statementAPI.bind(st)) { (freeS, func) =>
+          freeS.flatMap(boundSt => func(boundSt))
+        }
+
+    for {
+      _ <- log.debug("# Preparing insert query")
+      preparedStatement <- sessionAPI.prepare("INSERT INTO users (id, name) VALUES (?, ?)")
+      _ <- log.debug("# Binding values")
+      boundStatement    <- bindValues(preparedStatement)
+      _ <- log.debug(s"# Executing insert query with id ${newUser.id}")
+      _                 <- sessionAPI.executeStatement(boundStatement)
+      _ <- log.debug("# Selecting previous inserted item")
+      user <- sessionAPI.executeWithMap(
+        s"SELECT id, name FROM users WHERE id = ?",
+        Map("id" -> newUser.id)) map { rs =>
+        val row = rs.one()
+        User(row.getUUID(0), row.getString(1))
+      }
+      _ <- log.debug(s"# Fetched item $user")
+      _ <- log.debug(s"# Closing connection")
+      _ <- sessionAPI.close
+    } yield user
+
+  }
 
   def close[F[_]](implicit clusterAPI: ClusterAPI[F]): FreeS[F, Unit] = clusterAPI.close
 
-  val uuid = UUID.randomUUID()
-
-  implicit def instance[F[_]]: FreeSLift[F, monix.eval.Task] = new FreeSLift[F, monix.eval.Task] {
-    def liftFSPar[A](ga: monix.eval.Task[A]): FreeS.Par[F, A] = ga.liftFSPar[F]
-  }
-
-  def program[F[_]](implicit app: CassandraApp[F]): FreeS[F, User] = {
-
-    implicit val s = app.sessionAPI
-
-    val insertUserTask: FreeS[F, ResultSet] = cql"INSERT INTO users (id, name) VALUES ($uuid, 'Username');".asResultSet[F]
-    val getUserTask: FreeS[F, ResultSet]= cql"SELECT id, name FROM users WHERE id = $uuid".asResultSet[F]
-
-    for {
-      _ <- app.log.debug(s"# Executing insert query with id $uuid")
-      newUser <- insertUserTask
-      _ <- app.log.debug("# Selecting previous inserted item")
-      userResultSet <- getUserTask
-      user = {
-        val userRow = userResultSet.one()
-        User(userRow.getUUID(0), userRow.getString(1))
-      }
-      _ <- app.log.debug(s"# Fetched item: $user")
-      _ <- app.log.debug(s"# Closing connection")
-      _ <- app.sessionAPI.close
-    } yield user
-  }
+  implicit val executionContext: Scheduler = Scheduler.Implicits.global
 
   val beforeTask: MonixTask[Session] = connect[ClusterAPI.Op].interpret[MonixTask]
   implicit val session: Session = Await.result(beforeTask.runAsync, Duration.Inf)
@@ -86,8 +89,5 @@ object StringInterpolator extends App {
 
   val afterTask: MonixTask[Unit] = close[ClusterAPI.Op].interpret[MonixTask]
   Await.result(afterTask.runAsync, Duration.Inf)
-
-  Await.result(closeSession[SessionAPI.Op].interpret[MonixTask].runAsync, Duration.Inf)
-  Await.result(close[ClusterAPI.Op].interpret[MonixTask].runAsync, Duration.Inf)
 
 }
