+++ basic/app/src/main/scala/freestyle/cassandra/sample/Modules.scala
