+++ basic/build.sbt
@@ -1,4 +1,4 @@
-name := "freestyle-cassandra-low-level-api"
+val appName = "freestyle-cassandra-examples"
 
 scalaVersion := "2.12.3"
 
@@ -7,11 +7,28 @@ resolvers ++= Seq(
   Resolver.sonatypeRepo("snapshots"),
   "jitpack" at "https://jitpack.io")
 
-addCompilerPlugin("org.scalameta" % "paradise" % "3.0.0-M10" cross CrossVersion.full)
+val appPlugins = addCompilerPlugin("org.scalameta" % "paradise" % "3.0.0-M10" cross CrossVersion.full)
 
-libraryDependencies ++= Seq(
-  "io.frees" %% "freestyle-monix"       % "0.3.1",
-  "io.frees" %% "freestyle-async-monix" % "0.3.1",
-  "io.frees" %% "freestyle-logging"     % "0.3.1",
-  "io.frees" %% "frees-cassandra-core"  % "0.0.1"
+val dependencies = Seq(
+  "io.frees" %% "frees-monix"       % "0.4.1",
+  "io.frees" %% "frees-async"       % "0.4.1",
+  "io.frees" %% "frees-logging"     % "0.4.1",
+  "io.frees" %% "frees-async-cats-effect" % "0.4.1",
+  "io.frees" %% "frees-cassandra-core"  % "0.0.2"
 )
+
+lazy val model = (project in file("model"))
+  .settings(name := "model")
+  .settings(libraryDependencies ++= dependencies)
+  .settings(appPlugins)
+
+lazy val examples = (project in file("examples"))
+  .settings(name := "examples")
+  .settings(libraryDependencies ++= dependencies)
+  .settings(appPlugins)
+  .dependsOn(model)
+
+lazy val root = (project in file("."))
+  .settings(name := appName)
+  .aggregate(examples, model)
+  .dependsOn(examples, model)
