+++ basic/model/src/main/scala/freestyle/cassandra/sample/Model.scala
@@ -16,8 +16,15 @@
 
 package freestyle.cassandra.sample
 
+import freestyle.cassandra.query.interpolator.MacroInterpolator.SchemaFileInterpolator
+
 object Model {
 
   case class User(id: java.util.UUID, name: String)
+
+  @SchemaFileInterpolator("/schema.cql")
+  trait DummySchemaInterpolator
+
+
 
 }
